refactor(CorrelationView): :art: Add tests and refactor Correlation building
    related to issue #11
