+++ src/main.ts
@@ -19,6 +19,9 @@ import {
 	splitAndTrim,
 	stringToNullOrUndefined,
 } from "./utils";
+import {
+	buildAllCorrelations,
+} from "./correlationUtils";
 import { getPearsonCorrelation, getPointBiserialCorrelation } from "./analyses";
 import CorrelationView from "./CorrelationView";
 import { openView } from "obsidian-community-lib";
@@ -27,11 +30,13 @@ export default class DataAnalysisPlugin extends Plugin {
 	settings: Settings;
 	index: {
 		data: { [field: string]: DataType }[];
+		dataMap: {[field: string]: DataType[]};
 		corrs: Correlations;
 		minDate: DateTime;
 		maxDate: DateTime;
 	} = {
 		data: undefined,
+		dataMap: undefined,
 		corrs: undefined,
 		minDate: undefined,
 		maxDate: undefined,
@@ -48,7 +53,7 @@ export default class DataAnalysisPlugin extends Plugin {
 		const onAPIReady = async (api: DataviewApi) => {
 			this.app.workspace.onLayoutReady(async () => {
 				await this.refreshIndex(api);
-				this.index.corrs = this.buildAllCorrelations();
+				this.index.corrs = buildAllCorrelations(this.index.data, this.settings.fieldsToCheck);
 			});
 		};
 
@@ -90,7 +95,11 @@ export default class DataAnalysisPlugin extends Plugin {
 		this.addCommand({
 			id: "builds-corrs",
 			name: "Build Correlations",
-			callback: async () => console.log(this.buildAllCorrelations()),
+			callback: async () => {
+				const corrs = buildAllCorrelations(this.index.data, this.settings.fieldsToCheck);
+				// TODO: don't we want to set the index.corrs to the value here?
+				console.log(corrs);
+			},
 		});
 
 		this.addCommand({
@@ -271,130 +280,6 @@ export default class DataAnalysisPlugin extends Plugin {
 		const type = this.inferType(xs);
 		return xs.map((x) => x ?? (type === "number" ? 0 : "N/A"));
 	}
-
-	buildAllCorrelations() {
-		const { data } = this.index;
-		const { fieldsToCheck } = this.settings;
-		const corrs: Correlations = {};
-
-		for (const fA of fieldsToCheck) {
-			corrs[fA] = {};
-			const vA = data.map((d) => d[fA]);
-			const tA = this.inferType(vA);
-			for (const fB of fieldsToCheck) {
-				if (fA === fB) continue;
-				const vB = data.map((d) => d[fB]);
-				const tB = this.inferType(vB);
-
-				if (tA === "number" && tB === "number") {
-					if (corrs[fB]?.[fA]) continue;
-					const [oA, oB] = arrayOverlap(vA, vB);
-					const corr = getPearsonCorrelation(oA, oB);
-					corrs[fA][fB] = corr ? { corr, n: oA.length } : null;
-				} else if (tA === "number" && tB === "string") {
-					const oA = vA.filter((a) => a);
-					const oB = vB
-						.filter((b, i) => vA[i] !== undefined)
-						.map((b) => b ?? 0);
-
-					const uniqueStrs = [...new Set(oB)].filter(
-						(str) => typeof str === "string"
-					);
-					uniqueStrs.forEach((subF) => {
-						const subA = oA;
-						const subB = oB.map((b) => (b === subF ? 1 : 0));
-
-						const corr = getPointBiserialCorrelation(subB, subA);
-						corrs[fA][fB + "." + subF] = corr
-							? { corr, n: subA.length }
-							: null;
-					});
-				} else if (tA === "number" && tB === "object") {
-					const oA = vA.filter((a) => a);
-					const oB = (vB as string[][]).filter(
-						(b, i) => vA[i] !== undefined
-					);
-
-					const uniqueStrs = [...new Set(oB.flat())].filter(
-						(str) => typeof str === "string"
-					);
-					uniqueStrs.forEach((subF) => {
-						const subA = oA;
-						const subB = oB.map((b) =>
-							b && b.includes(subF) ? 1 : 0
-						);
-						const corr = getPointBiserialCorrelation(subB, subA);
-						corrs[fA][fB + "." + subF] = corr
-							? { corr, n: subA.length }
-							: null;
-					});
-				} else if (tA === "string" && tB === "string") {
-					corrs[fA][fB] = null;
-				} else if (tA === "string" && tB === "number") {
-				} else if (tA === "string" && tB === "object") {
-					corrs[fA][fB] = null;
-				} else if (tA === "object" && tB === "object") {
-					corrs[fA][fB] = null;
-				} else if (tA === "object" && tB === "number") {
-				} else if (tA === "object" && tB === "string") {
-					corrs[fA][fB] = null;
-				}
-			}
-		}
-		console.log({ corrs });
-		return corrs;
-	}
-
-	// getAllCorrsForField(fieldA: string) {
-	// 	const { data } = this.index;
-	// 	const { fieldsToCheck } = this.settings;
-	// 	const correlations = {};
-
-	// 	const fieldsForA = this.allUniqueValuesForField(fieldA);
-
-	// 	fieldsToCheck.forEach((fieldB) => {
-	// 		const fieldsForB = this.allUniqueValuesForField(fieldB);
-
-	// 		const valsInCommon: {
-	// 			[fieldA: string]: {
-	// 				[fieldB: string]: [
-	// 					string | number | string[],
-	// 					string | number | string[]
-	// 				];
-	// 			};
-	// 		} = {};
-
-	// 		if (!valsInCommon.hasOwnProperty(fieldA)) {
-	// 			valsInCommon[fieldA] = {};
-	// 		}
-	// 		const valsA = data.map((d) => d[fieldA]);
-	// 		const valsB = data.map((d) => d[fieldB]);
-
-	// 		valsA.forEach((valA) => {
-	// 			if (typeof valA === "string") {
-	// 				// valsInCommon[fieldA + valA] = valA;
-	// 			}
-	// 		});
-	// 		valsInCommon[fieldA][fieldB] = [valsA, valsB];
-
-	// 		// for (const fieldA in valsInCommon) {
-	// 		// 	for (const fieldB in valsInCommon) {
-	// 		// 		const [valA, valB] = valsInCommon[fieldA][fieldB];
-	// 		// 		[valA, valB].forEach((val, i) => {
-	// 		// 			const arr = makeArr(val);
-	// 		// 			if (typeof arr[0] === "string") {
-	// 		// 				valsInCommon[fieldA][fieldB][i] = arr.map((x) => {
-	// 		// 					if (i === 0)
-	// 		// 						return fieldsForA.includes(x) ? 1 : 0;
-	// 		// 					else return fieldsForB.includes(x) ? 1 : 0;
-	// 		// 				});
-	// 		// 			}
-	// 		// 		});
-	// 		// 	}
-	// 		// }
-	// 		console.log(valsInCommon);
-	// 	});
-	// }
 
 	async createJSDF() {
 		const { settings } = this;
